requirements.txt
streamlit==1.33.0
google-generativeai==0.5.2
python-dotenv==1.0.1
langchain==0.1.16
PyPDF2==3.0.1
chromadb==0.4.24
faiss-cpu==1.8.0
weaviate-client==4.5.5
langchain_weaviate==4.5.5
langchain_google_genai==1.0.2
chardet==5.2.0
pypdf==4.2.0
langchain-community==0.0.34
